test: align mitm webhook tests with delivery boundaries
